moved simplifyDriver back to src/driver
